requirements.txt
aiohttp>=3.8.0,<4.0.0
aiosignal>=1.2.0,<2.0.0
altair>=4.2.0,<6.0.0
annotated-types>=0.5.0
anyio>=3.6.1,<4.0.0
appnope>=0.1.3
argon2-cffi>=21.3.0
argon2-cffi-bindings>=21.2.0
asttokens>=2.0.8
async-timeout>=4.0.0
attrs>=22.1.0
backcall>=0.2.0
beautifulsoup4>=4.11.1
bleach>=5.0.1
blinker>=1.6.0
cachetools>=5.3.0
certifi>=2023.7.22
cffi>=1.15.1
charset-normalizer>=3.0.0
click>=8.0.0
dataclasses-json>=0.5.7
debugpy>=1.6.3
decorator>=5.1.1
defusedxml>=0.7.1
distro>=1.8.0
duckdb>=0.8.1
entrypoints>=0.4
executing>=1.1.1
fastjsonschema>=2.16.2
filelock>=3.9.0
frozenlist>=1.3.3
fsspec>=2023.1.0
gitdb>=4.0.10
GitPython>=3.1.30
groq>=0.16.0
h11>=0.14.0
httpcore>=0.17.0
httpx>=0.24.0
huggingface-hub>=0.16.4
idna>=3.4
importlib-metadata>=6.0.0
ipykernel>=6.16.1
ipython>=8.5.0
ipython-genutils>=0.2.0
ipywidgets>=8.0.2
jedi>=0.18.1
Jinja2>=3.1.2
joblib>=1.3.0
jsonpatch>=1.33
jsonpointer>=2.4
jsonschema>=4.16.0
jupyter>=1.0.0
jupyter-console>=6.4.4
jupyter-server>=1.21.0
jupyter_client>=7.4.3
jupyter_core>=4.11.2
jupyterlab-pygments>=0.2.2
jupyterlab-widgets>=3.0.3
langchain>=0.0.300
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-groq>=0.0.1
langsmith>=0.0.87
markdown-it-py>=3.0.0
MarkupSafe>=2.1.1
marshmallow>=3.19.0
matplotlib-inline>=0.1.6
mdurl>=0.1.2
mistune>=2.0.4
mpmath>=1.3.0
multidict>=6.0.4
mypy-extensions>=1.0.0
nbclassic>=0.4.5
nbclient>=0.7.0
nbconvert>=7.2.2
nbformat>=5.7.0
nest-asyncio>=1.5.6
networkx>=3.0
nltk>=3.8.1
notebook>=6.5.1
notebook_shim>=0.2.0
numpy>=1.23.4
openai>=1.0.0
packaging>=23.0
pandas>=1.5.1
pandocfilters>=1.5.0
parso>=0.8.3
pexpect>=4.8.0
pickleshare>=0.7.5
pillow>=10.0.0
pinecone-client>=2.2.2
prometheus-client>=0.15.0
prompt-toolkit>=3.0.31
protobuf>=4.25.0
psutil>=5.9.3
ptyprocess>=0.7.0
pure-eval>=0.2.2
pyarrow>=12.0.0
pycparser>=2.21
pydantic>=2.0.0
pydantic_core>=2.0.0
pydeck>=0.8.0
Pygments>=2.13.0
pyparsing>=3.0.9
pyrsistent>=0.18.1
python-dateutil>=2.8.2
pytz>=2022.5
PyYAML>=6.0
pyzmq>=24.0.1
qtconsole>=5.3.2
QtPy>=2.2.1
regex>=2023.0.0
requests>=2.28.0
rich>=13.0.0
safetensors>=0.4.0
scikit-learn>=1.3.0
scipy>=1.10.0
Send2Trash>=1.8.0
sentence-transformers>=2.2.2
sentencepiece>=0.1.99
six>=1.16.0
smmap>=5.0.0
sniffio>=1.3.0
soupsieve>=2.3.2
SQLAlchemy>=2.0.0
sqlparse>=0.4.4
stack-data>=0.5.1
streamlit>=1.25.0
sympy>=1.12
tenacity>=8.2.0
terminado>=0.16.0
threadpoolctl>=3.1.0
tiktoken>=0.4.0
tinycss2>=1.2.1
tokenizers>=0.13.0
toml>=0.10.2
toolz>=0.12.0
torch>=2.0.0
tornado>=6.2
tqdm>=4.65.0
traitlets>=5.5.0
transformers>=4.30.0
typing-inspect>=0.8.0
typing_extensions>=4.5.0
tzlocal>=5.0.0
urllib3>=2.0.0
validators>=0.20.0
wcwidth>=0.2.5
webencodings>=0.5.1
websocket-client>=1.4.1
widgetsnbextension>=4.0.3
yarl>=1.8.0
zipp>=3.15.0
psycopg2>=2.9.9
python-dotenv==0.19.0